move hc to seperate files
