+++ src/block/mod.rs
@@ -59,7 +59,7 @@ const LAST_LITERALS: usize = 5;
 /// spec.
 const END_OFFSET: usize = LAST_LITERALS + 1;
 
-/// https://github.com/lz4/lz4/blob/dev/doc/lz4_Block_format.md#end-of-block-restrictions
+/// https://github.com/lz4/lz4/blob/dev/doc/lz4_Block_format.md#end-of-block-conditions
 /// Minimum length of a block
 ///
 /// MFLIMIT + 1 for the token.
@@ -68,18 +68,8 @@ const LZ4_MIN_LENGTH: usize = MFLIMIT + 1;
 const MAXD_LOG: usize = 16;
 const MAX_DISTANCE: usize = (1 << MAXD_LOG) - 1;
 
-#[allow(dead_code)]
-const MATCH_LENGTH_MASK: u32 = (1_u32 << 4) - 1; // 0b1111 / 15
-
 /// The minimum length of a duplicate
 const MINMATCH: usize = 4;
-
-#[allow(dead_code)]
-const FASTLOOP_SAFE_DISTANCE: usize = 64;
-
-/// Switch for the hashtable size byU16
-#[allow(dead_code)]
-static LZ4_64KLIMIT: usize = (64 * 1024) + (MFLIMIT - 1);
 
 /// An error representing invalid compressed data.
 #[derive(Debug)]
